Return an int

+++ 0x00-python_variable_annotations/2-floor.py
@@ -4,13 +4,13 @@ Module contains function that floors a float
 '''
 import math
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     '''Takes a float and returns the floor of the float
 
     Args:
         n (float): Float to be floored
 
     Returns:
-        float: The floored float
+        int: The floored float
     '''
     return math.floor(n)
